Guile and SDL changed places
Now the SDL thread is started from Guile. Before it was the SDL
thread which stated the Guile thread.

Also changed so that the main is now supposed to be a Guile file,
and load the c code as a shared object. Before the c code loaded
the guile code as is.