Scene Macros
All scenes have enough boilerplate to be annoying.

You have to define the module, import a bunch of the
same objects. Have a bunch of define-once stuff,
defining the actual scene object. Calling everything
else with that.

This commit solves all that except the import part.

Currenttly only implemented in scene3