New event system should now work!
New much simpler and better event system was actually rather easy to get
to work as it should. It still however needs a bit of cleanup, a few of
the objects now make incorrect assumptions about how the objects are.

There also is still quite a bit of need for cleaning up the code and
commenting everything.